requirements.txt
# local package
-e .

# external requirements
click==7.1.*
sphinx==4.1.*
coverage==5.5
awscli==1.20.*
mypy==0.910
black==21.7b0
python-dotenv==0.19.*
pytest==6.2.*
pillow==8.3.*
coloredlogs==15.0.*
pymongo[srv]==3.12.*
sqlalchemy==1.4.*
beautifulsoup4==4.9.*
pandas==1.3.*
requests==2.26.*
pydantic==1.8.*
tqdm==4.62.*
types-requests==2.25.*
wordcloud==1.8.*
pyspark[sql]==3.1.*
lxml==4.6.*
kaggle==1.5.*
spark-nlp==3.3.*
nltk==3.6.*